fix broken pre-rendering of the playbooks stored in localStorage
